updated file upload logic

+++ index.php
@@ -81,8 +81,12 @@
         $file_tmp = $_FILES['upload']['tmp_name'];
         $file_type = $_FILES['upload']['type'];
         if (!empty($file_name)) {
-            move_uploaded_file($file_tmp, $_GET['path'] . $file_name);
-            print("<script type=\"text/javascript\">alert('File uploaded successfully!')</script>");
+            if (is_file($file_name)) {
+                print("<script type=\"text/javascript\">alert('Unable to upload a file with a name that already exists.')</script>");
+            } else {
+                move_uploaded_file($file_tmp, $_GET['path'] . $file_name);
+                print("<script type=\"text/javascript\">alert('File uploaded successfully!')</script>");
+            }
         }
     }
